requirements.txt
flask==3.1.0
flask-session==0.8.0
flask-limiter==3.9.0
spotipy==2.26.0
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
cryptography==44.0.2
apscheduler==3.10.4
gunicorn==23.0.0
